[PATCH] Wallet Withdraw Done

+++ database/migrations/2023_01_15_124224_create_wallets_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
-            $table->float('wallet_amount')->default(0);
+            $table->integer('wallet_amount')->default(0);
             $table->string('wallet_method')->nullable();
             $table->string('wallet_name')->nullable();
             $table->string('wallet_rekening')->nullable();
